Delete CompilationErrorsTest now that compile-testing is powerful enough to assert everything it was asserting. Port the test methods into CompilationTest.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=152940215
